requirements.txt
absl-py==2.0.0
aiohttp==3.9.0
aiosignal==1.2.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dm-tree==0.1.8
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
flatbuffers==23.5.26
frozenlist==1.4.0
gast==0.4.0
google-auth==2.24.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.14.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.0.4
namex==0.0.7
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
Pillow==10.1.0
pip==23.3.1
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
Pygments==2.17.2
PyJWT==2.4.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scipy==1.11.3
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.42.0
win-inet-pton==1.1.0
wrapt==1.14.1
yarl==1.9.3
